requirements.txt
Flask==2.3.3
PyNaCl==1.5.0
qrcode==7.4.2
Pillow==10.0.1
imagehash==4.3.1
scikit-image==0.21.0
opencv-python==4.8.1.78
numpy==1.24.4